refactor(onboarding): use layout component for welcome page
