+++ src/pages/onboarding/magic-recovery-code/magic-recovery-code.tsx
